!!!MAJOR BUG FIX!!!
When the TaskManager attempts to spawn a new worker, while also inside of an active "DB" lock (via Padre::Locker) the database needs to be disconnected at the moment of the thread spawn.

The code that reconnects to the database uses the non-existant method Padre::DB->begin_work instead of Padre::DB->begin, resulting in corrupt TaskManager state and a failed unlock.

git-svn-id: http://svn.perlide.org/padre/trunk/Padre@13823 bc93166f-e056-4eb9-8c59-0a0840647e1b