on_right_click => on_context_menu
A first step to go to a clean support of the context menu key (available on
Windows keyboards) and its alternatives (right click or Shift+F10) in the editor.
But on GTK at least, EVENT_CONTEXT_MENU does not seem to work in Scintilla. So
we still use EVENT_RIGHT_CLICK to trigger the menu.


git-svn-id: http://svn.perlide.org/padre/trunk/Padre@15822 bc93166f-e056-4eb9-8c59-0a0840647e1b